Extracted queue polling strategy
In order to be able to customize the order and manner in which Shoryuken
polls SQS queues, the previous hard-coded strategy was extracted into a
separate class.

This meant that some of the method names were altered to be more generic
(for example rebalance_queue_weight! became messages_received),
otherwise this was a direct extraction.