Add visibility_timeout attribute to Shoryuken::Message
Ideally we should be able to query a message's visibility timeout using
`self.data.visibility_timeout`. However, [aws-sdk-sqs] doesn't seem to
provide such API support.

This is the simplest solution to have Shoryuken::Message to manage its
own visibility timeout.

It's also worth noting that the visibility timeout, whenever set, is
relevant to the time the `ChangeMessageVisibility` request is issued.
This can be found in [aws-doc]. This requires us to manually compute the
absolute time when the visibility times out using

```
Time.now + @visibility_timeout
```

[aws-sdk-sqs]:
https://github.com/aws/aws-sdk-ruby/blob/master/gems/aws-sdk-sqs/lib/aws-sdk-sqs/message.rb
[aws-doc]:
https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-visibility-timeout.html#changing-message-visibility-timeout