Change `Manager#assign` to accept an array of messages
Instead of creating a Promise to process a single message, this will
chain together a series of Promises so that all the messages are
processed serially. If an array with only a single message is passed,
this will behave the same as it did before.

This change is in preparation for a future change where we will pass an
array of messages with a common message group ID.