Remove the 1 message fetch limit for FIFO queues
This was originally introduced to ensure that messages in a FIFO queue
weren't processed out of order. However, this was overly limiting
because while we want to process messages with a common message group ID
in order, we can process messages with different message group IDs out
of order[1]. With the changes in previous commits to ensure that we process
messages with the same message group ID in order, we can now fetch more
messages at a time, improving throughput and reducing the number of
calls to AWS that we make.

[1] https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/using-messagegroupid-property.html