Remove `concurrency` limit when fetching messages from a FIFO queue
Because we can now assign multiple messages to a single processor to
handle serially, limiting the amount of messages that we fetch to the
number of available processors isn't always a good idea (at least with
regard to FIFO queues). For example, if all of the messages currently in
a FIFO queue have the same message group ID, it's inefficient to fetch
each message individually when instead we could fetch as many as AWS
will let us and process them all serially.

A consequence of this change is that in any given fetch, we could
receive a group of messages distributed across some number of message
groups (say, 5) while there is a smaller number of processors available
(say, 3). In this case we would end up creating more Promises than we
theoretically have resources for. The worst case scenario would be
fetching 10 messages, each with their own message group ID, when we only
have 1 available processor. It is my understanding that the underlying
executor uses a queue to store pending work so I don't _think_
overprovisioning is the end of the world but I could be wrong.