Add appraisals for Rails 4.2, 5.2, 6.1, and 6.2
The .gemfile files generated by `appraisal install` are checked-in to
git, but the .gemfile.lock files are ignored, matching appraisal's
recommendation.

The list of Rails versions was determined by Rails' [maintenance
policy][0], including everything in "severe security issues". In
addition, support for Rails 4.2 is included, despite it being EOL'd by
Rails. Removal of support for Rails 4.2 in shoryuken would be considered
a breaking change, so we'll keep support for now.

[0]: https://guides.rubyonrails.org/maintenance_policy.html