Execute Rails suite against multiple versions of Rails in CI
On Travis, we'll run the Rails test suite against all supported versions
of Rails. In order to avoid running too many jobs in Travis, we only run
each Rails version against a single Ruby version. The Ruby version for
each suite was determined by latest stable version compatible with the
Rails version. The [compatibility matrix on the fastruby blog][0] is a
good reference for this.

[0]: https://www.fastruby.io/blog/ruby/rails/versions/compatibility-table.html