Replace ActiveJob double with real instance in specs
This'll help ensure that the ActiveJob adapters continue to work with
all supporter versions of ActiveJob.

The `require` order is unfortunately very sensitive here: it's important
to force the `:autoload` of `ActiveJob::QueueAdapters` (for example, by
defining a subclass of `ActiveJob::Base`) before defining our own
`ActiveJob::QueueAdapters::Shoryuken{Adapter,ConcurrentSendAdapter}`.
Otherwise, we end up in a situation where class methods on
`ActiveJob::QueueAdapters` are not defined.