Compute batch payload size more accurately
The previous algorithm computed the size of a batch of messages using
the Ruby literal representation of the messages, which overestimates the
true size of the messages as SQS computes it. While overestimating is
relatively benign, it does mean that we can end up making more SQS API
calls than necessary.