Send messages in fewer batches
The previous algorithm considered each `max_batch_size`-sized slice of
the messages in order. For each slice, it'd subdivide the slice until
the batches were small enough to send.

To minimize the number of batches we send, however, it makes more sense
to take as many messages as we can from the beginning of the list of
messages, and just do that repeatedly until we've sent them all.