Update `dispatch` hook to have utilization metrics
I'd like to export metrics from Shoryuken so that we have metrics available for the number of worker threads busy at any given time. The logical thing to do is to put the metrics in the `dispatch` lifecycle hook, but right now they don't have that information available.

I'm suggesting that we add that information to the arguments that are passed to the hook, so that it is available within the hook.